Bottle work started and test fix

+++ Server/src/main/java/fr/univ_smb/isc/m2/repository/ColorRepository.java
@@ -3,6 +3,6 @@ package fr.univ_smb.isc.m2.repository;
 import fr.univ_smb.isc.m2.models.Color;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ColorRepository extends JpaRepository<Color, Long> {
+public interface ColorRepository extends JpaRepository<Color, Integer> {
 }
